Fix incorrect coloring appearing periodically with CUDA strategy
Fix red/green coloring appearing periodically with some choices of cell
dimensions when using the CUDA-based board update strategy.

When allocating arrays to pass to the CUDA kernel, copy the color values
to their device array even though the kernel doesn't read them.
Otherwise for some reason, writes intended to set the new cells
periodically appear in the colors array, despite the index written being
within the bounds of the new cells array (determined via debug
printouts).

Costs about 8% in framerate.

Add a test that colors are correctly set when updating a board with an
OpenGL draw state provided. Does not reproduce the problem but is a good
test to confirm that update strategies do this.